Backend/requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
curl_cffi==0.13.0
Flask==3.1.2
flask-cors==6.0.1
frozendict==2.4.7
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multitasking==0.0.12
mysql-connector-python==9.5.0
numpy==2.4.4
pandas==3.0.2
peewee==4.0.4
platformdirs==4.9.4
protobuf==7.34.1
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.5.0
websockets==16.0
Werkzeug==3.1.3
yfinance==1.2.0